backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = {extras = ["standard"], version = "^0.118.0"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^3.2.0"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
pydantic-settings = "^2.11.0"
pydicom = "^3.0.1"
pynetdicom = "^3.0.4"
pillow = "^11.3.0"
numpy = "^2.3.3"
pandas = "^2.3.3"
python-dateutil = "^2.9.0.post0"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
